# AK+LibJS: Performance improvements for resolving large rope strings