--- a/AK/StringBuilder.h
+++ b/AK/StringBuilder.h
@@ -18,6 +18,7 @@ class StringBuilder {
 public:
     static constexpr size_t inline_capacity = 256;
 
+    using Buffer = Detail::ByteBuffer<inline_capacity>;
     using OutputType = ByteString;
 
     static ErrorOr<StringBuilder> create(size_t initial_capacity = inline_capacity);
@@ -61,8 +62,8 @@ public:
 
     [[nodiscard]] ByteString to_byte_string() const;
 
-    [[nodiscard]] String to_string_without_validation() const;
-    ErrorOr<String> to_string() const;
+    [[nodiscard]] String to_string_without_validation();
+    ErrorOr<String> to_string();
 
     [[nodiscard]] FlyString to_fly_string_without_validation() const;
     ErrorOr<FlyString> to_fly_string() const;
@@ -95,12 +96,16 @@ public:
         return {};
     }
 
+    Optional<Buffer::OutlineBuffer> leak_buffer_for_string_construction(Badge<Detail::StringData>);
+
 private:
+    explicit StringBuilder(Buffer);
+
     ErrorOr<void> will_append(size_t);
     u8* data();
     u8 const* data() const;
 
-    Detail::ByteBuffer<inline_capacity> m_buffer;
+    Buffer m_buffer;
 };
 
 }