This makes 2 changes to improve rope string resolution:
1. Allow `StringBuilder` to transfer ownership of its string buffer to `String`, so we don't have to re-allocate and copy the string data
2. Allocate the resolved rope's `StringBuilder` capacity up front

For a ballpark idea of how we perform relative to other rope string implementations, I compared us to [fatal's rope string](https://github.com/facebook/fatal/blob/main/fatal/string/rope.h).

<details>
<summary>Benchmark code</summary><p>

```c++
static auto vm = JS::VM::create().release_value();

static auto create_libjs_rope()
{
    auto string = JS::PrimitiveString::create(vm, MUST(String::repeated("abcdef"_string, 200)));

    auto rope = JS::PrimitiveString::create(vm, string, string);
    for (size_t i = 0; i < 1'000'000; ++i)
        rope = JS::PrimitiveString::create(vm, rope, string);

    return rope;
}

static auto create_fatal_rope()
{
    auto string = ByteString::repeated("abcdef"sv, 200);

    auto rope = fatal::rope(string.characters(), string.characters());
    for (size_t i = 0; i < 1'000'000; ++i)
        rope.append(string.characters());

    return rope;
}

static auto libjs_rope = create_libjs_rope();
static auto fatal_rope = create_fatal_rope();

BENCHMARK_CASE(bench_libjs)
{
    auto resolved = libjs_rope->utf8_string();
    dbgln("libjs rope length = {}", resolved.bytes_as_string_view().length());
}

BENCHMARK_CASE(bench_fatal)
{
    std::string resolved = fatal_rope.to_string();
    dbgln("fatal rope length = {}", resolved.size());
}
```
</p></details>

| | Resolve Time |
|-|-|
| LibJS (before) | 2.07s |
| LibJS (after) | 0.466s |
| fatal | 0.478s |

The savings here amount to about ~0.5s for pre-allocating the `StringBuilder` and ~1.0s for handing the builder's buffer off without reallocation.